+++ kafka-impl/conf/kop_standalone.conf
@@ -71,6 +71,16 @@ offsetsTopicNumPartitions=8
 # Maximum number of entries that are read from cursor once per time
 maxReadEntriesNum=1
 
+# The format of an entry. The default value is pulsar.
+# Optional values: [pulsar]
+#
+# pulsar:
+#   When KoP receives messages from kafka producer, it will serialize these messages to
+#   the format so that pulsar consumer can read directly.
+#   When KoP sends entries to kafka consumer, it will treat each entry as pulsar's
+#   format and deserialize each entry to kafka's format.
+entry.format=pulsar
+
 ### --- KoP SSL configs--- ###
 
 # Kafka ssl configuration map with: SSL_PROTOCOL_CONFIG = ssl.protocol
