+++ kafka-impl/src/main/java/io/streamnative/pulsar/handlers/kop/utils/ByteBufUtils.java
@@ -1,0 +1,65 @@
+/**
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.streamnative.pulsar.handlers.kop.utils;
+
+import static java.nio.charset.StandardCharsets.UTF_8;
+
+import io.netty.buffer.ByteBuf;
+import java.nio.ByteBuffer;
+import java.util.Base64;
+import org.apache.pulsar.common.api.proto.PulsarApi;
+
+
+/**
+ * Utils for ByteBuf operations.
+ */
+public class ByteBufUtils {
+
+    public static ByteBuffer getKeyByteBuffer(PulsarApi.SingleMessageMetadata messageMetadata) {
+        if (messageMetadata.hasOrderingKey()) {
+            return messageMetadata.getOrderingKey().asReadOnlyByteBuffer();
+        }
+
+        String key = messageMetadata.getPartitionKey();
+        if (messageMetadata.hasPartitionKeyB64Encoded()) {
+            return ByteBuffer.wrap(Base64.getDecoder().decode(key));
+        } else {
+            // for Base64 not encoded string, convert to UTF_8 chars
+            return ByteBuffer.wrap(key.getBytes(UTF_8));
+        }
+    }
+
+    public static ByteBuffer getKeyByteBuffer(PulsarApi.MessageMetadata messageMetadata) {
+        if (messageMetadata.hasOrderingKey()) {
+            return messageMetadata.getOrderingKey().asReadOnlyByteBuffer();
+        }
+
+        String key = messageMetadata.getPartitionKey();
+        if (messageMetadata.hasPartitionKeyB64Encoded()) {
+            return ByteBuffer.wrap(Base64.getDecoder().decode(key));
+        } else {
+            // for Base64 not encoded string, convert to UTF_8 chars
+            return ByteBuffer.wrap(key.getBytes(UTF_8));
+        }
+    }
+
+    public static ByteBuffer getNioBuffer(ByteBuf buffer) {
+        if (buffer.isDirect()) {
+            return buffer.nioBuffer();
+        }
+        final byte[] bytes = new byte[buffer.readableBytes()];
+        buffer.getBytes(buffer.readerIndex(), bytes);
+        return ByteBuffer.wrap(bytes);
+    }
+}
